refactor: 리포지토리 패키지 경로 수정 및 UseCase 명명 규칙 통일
- `data` 모듈 내 리포지토리 구현체들의 패키지명을 `repotisoryimpl` 및 `repositoryimpl`에서 `repository`로 변경하고 이동
- `ObserveUseCase`를 `ObserveMindmapNodeUseCase`로 이름을 변경하여 목적을 명확히 함
- 변경된 클래스명 및 패키지 경로에 맞춰 DI 설정(`UseCaseModule`, `RepositoryModule`) 및 `MindMapViewModel` 수정
- `.vscode/settings.json` 추가 (자바 빌드 구성 자동 업데이트 설정)
