+++ data/src/main/java/com/hihihihi/data/di/RepositoryModule.kt
@@ -11,22 +11,22 @@ import com.hihihihi.data.remote.datasource.NaverDataSource
 import com.hihihihi.data.remote.datasource.QuoteRemoteDataSource
 import com.hihihihi.data.remote.datasource.SearchRemoteDataSource
 import com.hihihihi.data.remote.datasource.UserBookRemoteDataSource
-import com.hihihihi.data.repotisoryimpl.AuthRepositoryImpl
-import com.hihihihi.data.repotisoryimpl.KakaoAuthRepositoryImpl
-import com.hihihihi.data.repotisoryimpl.NaverAuthRepositoryImpl
+import com.hihihihi.data.repository.AuthRepositoryImpl
+import com.hihihihi.data.repository.KakaoAuthRepositoryImpl
+import com.hihihihi.data.repository.NaverAuthRepositoryImpl
 import com.hihihihi.data.remote.datasource.UserRemoteDataSource
-import com.hihihihi.data.repositoryimpl.SearchRepositoryImpl
-import com.hihihihi.data.repotisoryimpl.DailyReadPageRepositoryImpl
-import com.hihihihi.data.repotisoryimpl.HistoryRepositoryImpl
-import com.hihihihi.data.repotisoryimpl.MindmapNodeRepositoryImpl
-import com.hihihihi.data.repotisoryimpl.MindmapRepositoryImpl
-import com.hihihihi.data.repotisoryimpl.QuoteRepositoryImpl
-import com.hihihihi.data.repotisoryimpl.UserBookRepositoryImpl
-import com.hihihihi.data.repotisoryimpl.UserPreferencesRepositoryImpl
+import com.hihihihi.data.repository.SearchRepositoryImpl
+import com.hihihihi.data.repository.DailyReadPageRepositoryImpl
+import com.hihihihi.data.repository.HistoryRepositoryImpl
+import com.hihihihi.data.repository.MindmapNodeRepositoryImpl
+import com.hihihihi.data.repository.MindmapRepositoryImpl
+import com.hihihihi.data.repository.QuoteRepositoryImpl
+import com.hihihihi.data.repository.UserBookRepositoryImpl
+import com.hihihihi.data.repository.UserPreferencesRepositoryImpl
 import com.hihihihi.domain.repository.AuthRepository
 import com.hihihihi.domain.repository.KakaoAuthRepository
 import com.hihihihi.domain.repository.NaverAuthRepository
-import com.hihihihi.data.repotisoryimpl.UserRepositoryImpl
+import com.hihihihi.data.repository.UserRepositoryImpl
 import com.hihihihi.domain.repository.DailyReadPageRepository
 import com.hihihihi.domain.repository.HistoryRepository
 import com.hihihihi.domain.repository.MindmapNodeRepository
