+++ domain/src/main/java/com/hihihihi/domain/usecase/mindmapnode/ObserveMindmapNodeUseCase.kt
@@ -3,7 +3,7 @@ package com.hihihihi.domain.usecase.mindmapnode
 import com.hihihihi.domain.repository.MindmapNodeRepository
 import javax.inject.Inject
 
-class ObserveUseCase @Inject constructor(
+class ObserveMindmapNodeUseCase @Inject constructor(
     private val repository: MindmapNodeRepository
 ) {
     operator fun invoke(mindmapId: String) = repository.observe(mindmapId)
